+++ VBWEB/frontend/src/App.jsx
@@ -8,7 +8,6 @@ import LoginPage from './pages/LoginPage';
 import SignupPage from './pages/SignupPage';
 import ProfilePage from './pages/ProfilePage';
 import CourtInfoPage from './pages/CourtInfoPage';
-import UserInfoForm from './pages/UserInfoForm';
 import './App.css';
 
 function App() {
@@ -22,7 +21,6 @@ function App() {
           <Route path="/signup" element={<SignupPage />} />
           <Route path="/profile" element={<ProfilePage />} />
           <Route path="/court-info" element={<CourtInfoPage />} />
-          <Route path="/user-info" element={<UserInfoForm />} />
         </Routes>
       </main>
       <Footer />
