+++ VBWEB/frontend/src/pages/OldProfilePage.css
@@ -1,0 +1,22 @@
+.profile-page {
+  max-width: 600px;
+  margin: 4rem auto;
+  padding: 2rem;
+  background: #fff;
+  border-radius: 8px;
+  box-shadow: 0 4px 8px rgba(0,0,0,0.1);
+  text-align: center;
+  font-family: Arial, sans-serif;
+}
+
+.profile-page h2 {
+  color: #333;
+  margin-bottom: 1.5rem;
+  font-size: 2rem;
+}
+
+.profile-page p {
+  color: #555;
+  font-size: 1rem;
+  line-height: 1.6;
+}
