+++ VBWEB/frontend/src/pages/OldProfilePage.jsx
@@ -1,0 +1,16 @@
+// src/pages/ProfilePage.js
+import React from 'react';
+import './ProfilePage.css';
+
+function ProfilePage() {
+  return (
+    <div className="profile-page">
+      <h2>Your Profile</h2>
+      {/* Display user details and settings */}
+      <p>Welcome, [User Name]!</p>
+      {/* You can add forms or details to allow profile editing */}
+    </div>
+  );
+}
+
+export default ProfilePage;
