+++ VBWEB/frontend/src/pages/ProfilePage.css
@@ -1,22 +1,23 @@
-.profile-page {
+.user-info-form {
   max-width: 600px;
-  margin: 4rem auto;
+  margin: 2rem auto;
   padding: 2rem;
   background: #fff;
   border-radius: 8px;
   box-shadow: 0 4px 8px rgba(0,0,0,0.1);
-  text-align: center;
-  font-family: Arial, sans-serif;
 }
 
-.profile-page h2 {
-  color: #333;
+.user-info-form h2 {
   margin-bottom: 1.5rem;
-  font-size: 2rem;
 }
 
-.profile-page p {
-  color: #555;
-  font-size: 1rem;
-  line-height: 1.6;
+.user-info-form form label {
+  display: block;
+  margin-bottom: 1rem;
+}
+
+.user-info-form form input {
+  width: 100%;
+  padding: 0.5rem;
+  margin-top: 0.25rem;
 }
